refactor : header 중복된 내용 한쪽으로 정리
정상 / 오류 header를 나눠서 담당하도록 해서, 기존에 쓰던 필요없는 클래스를 삭제
